requirements.txt

streamlit
psycopg2-binary
python-docx
pytesseract
PyMuPDF
Pillow
